Make a window and put grid in
The window is display flex so its width equals window width, its aspect ratio is 1/1 so height is equal to width, i.e. square, and flex-wrap applied so the individual grids do not overflow, the individual grids' width is set at 6.25% so that 16 grids fit in one line and will resize with the container.